lib/alg-yescrypt-platform.c: Fix -Werror=sign-conversion.
In 894aee75433b4dc8d9724b126da6e79fa5f6814b we introduced some
changes, which show is this error when building with GCC v12.2.1.

+++ lib/alg-yescrypt-platform.c
@@ -47,7 +47,7 @@ static void *alloc_region(yescrypt_region_t *region, size_t size)
 	size_t new_size = size;
 	const size_t hugepage_mask = (size_t)HUGEPAGE_SIZE - 1;
 	if (size >= HUGEPAGE_THRESHOLD && size + hugepage_mask >= size) {
-		flags |= MAP_HUGETLB | MAP_HUGE_2MB;
+		flags |= (int)(MAP_HUGETLB | MAP_HUGE_2MB);
 /*
  * Linux's munmap() fails on MAP_HUGETLB mappings if size is not a multiple of
  * huge page size, so let's round up to huge page size here.
@@ -59,7 +59,7 @@ static void *alloc_region(yescrypt_region_t *region, size_t size)
 	if (base != MAP_FAILED) {
 		base_size = new_size;
 	} else if (flags & MAP_HUGETLB) {
-		flags &= ~(MAP_HUGETLB | MAP_HUGE_2MB);
+		flags &= ~(int)(MAP_HUGETLB | MAP_HUGE_2MB);
 		base = mmap(NULL, size, PROT_READ | PROT_WRITE, flags, -1, 0);
 	}
